pyproject.toml
[tool.poetry]
name = "code-snippets"
version = "0.1.0"
description = ""
authors = ["expiratory <expiratory@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.120.2"
sqlalchemy = "^2.0.44"
uvicorn = "^0.38.0"
python-dotenv = "^1.2.1"
psycopg2-binary = "^2.9.11"
asyncpg = "^0.29.0"
elasticsearch = "^8.12.0"
aio-pika = "^9.4.0"
redis = "^5.0.1"
aiohttp = "^3.13.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
email-validator = "^2.1.0"
slowapi = "^0.1.9"
authlib = "^1.6.5"
httpx = "^0.28.1"
itsdangerous = "^2.2.0"


[tool.poetry.group.dev.dependencies]
alembic = "^1.17.1"


[tool.poetry.group.prod.dependencies]
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
